CIV-17723 release clean up

+++ src/main/java/uk/gov/hmcts/reform/civil/handler/callback/camunda/dashboardnotifications/claimant/DefendantResponseWelshClaimantDashboardNotificationHandler.java
@@ -62,7 +62,8 @@ public class DefendantResponseWelshClaimantDashboardNotificationHandler extends 
 
     @Override
     protected void beforeRecordScenario(CaseData caseData, String authToken) {
-        if (caseData.getRespondent1ResponseDeadline().isBefore(LocalDateTime.now())) {
+        if (caseData.getRespondent1ResponseDeadline() != null
+            && caseData.getRespondent1ResponseDeadline().isBefore(LocalDateTime.now())) {
             dashboardNotificationService.deleteByNameAndReferenceAndCitizenRole(DJ_NOTIFICATION,
                     caseData.getCcdCaseReference().toString(), "CLAIMANT");
         }
